Add CheckBeamer() function. It will detect whether this tex file using documentclass beamer. If we are using beamer, it set g:Tex_DefaultTargetFormat='pdf', which will use xpdf for viewer. Otherwise we set it to dvi by default.
The problem is that this function only got called when we open the file. So if
you have two tabs with different documentclass. It can't change with different
tab. I tried to use autocmd BufEnter for this. It CAN detect and change the
g:Tex_DefaultTargetFormat correctly, but latex-suite doesn't response to the
change. Maybe later on I can figure this out. But so far this should be good
for me since I don't think I can work on the presentation and note at the same
time.