Mapping F4 to AutoMake()
I created the AutoMake function using Python. It will detect the included
files and automatic setting the makeprg for you. So it would very easy for
some simple exe tests and you don't need to edit Makefile every time.
So far, it only includes the cases for pthread, root and boost. If you need
anything else, it won't be hard to add.