Another improvement of automake
Getting correct included header and source code.
Define the compiler and flags with global variables.
Unfortunately with this command, g++ will compile every
source files. I haven't be able to find the way as smart
as Make. So this automake should only be useful for a small or test project. And this is what it is designed for.