AutoMake detects main()
It now can detect whether or not the target contains main().
If it is just a standalone code, it will compile to object file
with necessary includes directories. If it is a main, it will
compile all the code and make to executable.

However, I can't find the way to compile and link in one g++ command.
For example, main.cpp needs functions in test.cpp. While the test.cpp
have a object file which is ready to link. With g++ -o main main.cpp test.cpp
it will compile the test.cpp again. I think Make will seperate this into
two steps, compile main.cpp only and then link both into exe.
However, this can't be done with MakePrg in Vim. Running make
twice in one function, not a very good idea to me at this moment.
So, we still need to suffer the long compile time with this.....