Perform compile and link in two steps
Even though I complained about it would not be reasonable to do this. I
still tried to make it happen and it did!!

By sacrifying the length of the code, it can now do the following:
1. single source file, compile into an object file and put it into
the obj directory if available, otherwise just put it in local.
2. Single main file, compile and link to an executable
3. Mail file with multi sources, it cross check the timestamps of the
source and header file with the objects files. It would compile again if
needed and then linked all the obj files into one executable just like
normal make file. If no need for compile, it just perform the linking.

Even though the originally idea for this key mapping is just for me to
avoid adding thing to the Makefile. It seems way beyond my plan now..