Rescue =
Even though the toggle is so great, the original = is also quite useful for
our coding. So I made some modification to the toggle.
If it can't toggle word change, it will fall back to the default = and accept
motion keys. Unfortunately, I can't show the motions key in showcmd. But
it is show by echo now. It should work as well.