Add Locator "gl"
This is bundle for gl mapping in the vim, which show the location of current
file when the command is issued. I combined the pythonhelper and the locator
plugins from vim.org. locator is not perfect yet, especially for the C files.
I will need to work on this a little bit more