Quickfix result in powerline
Sometime quickfix can pickup a long list of output from make. I always need to
scroll to the end to find out whether the compiling is done successfully.
By taking advantage of the Makefile config and the powerline, we can make it
easier to use.

Example in Makefile:
XTDraw: XTDraw.o StopStyle.o
	$(LD) $(LDFLAGS) $^ $(LIBS) $(OutPutOpt) $@
	@echo "$@ done"

In case of successful compiling, the end of quickfix should have something
like "XXXXX done". This script is try to pick up the symbol from a global
variable g:qflist_done and indicate it on the powerline.