Improve file searching
The file searching in vim will go all the way up the /.
This could cause the vim to slow down significantly on those
public machines with different disk systems mounted.

Now use SearchFile to search upward until it reach the directory
named as the user's name. This should improve the startuptime
on those public machines.